One remaining explicit conversion that does affect code generation
- This conversion results in an extra SIGNEXTEND instruction that does not get optimized out. This makes the code more expensive by 12 GAS and makes gas tests fail.

+++ contracts/libraries/Oracle.sol
@@ -274,7 +274,7 @@ library Oracle {
             uint32 targetDelta = target - beforeOrAt.blockTimestamp;
             return (
                 beforeOrAt.tickCumulative +
-                    ((atOrAfter.tickCumulative - beforeOrAt.tickCumulative) / observationTimeDelta) *
+                    ((atOrAfter.tickCumulative - beforeOrAt.tickCumulative) / int56(uint56(observationTimeDelta))) *
                     int56(uint56(targetDelta)),
                 beforeOrAt.secondsPerLiquidityCumulativeX128 +
                     uint160(
